Changed classifier interface: result is a real value in [-1, 1] to express uncertainty. Refactoring: using "using" statement.

+++ Text_classifier/Text_classifier/Classification/Classifier.cs
@@ -12,7 +12,8 @@ namespace Text_classifier.Classification
         void Train(String text1, String text2);
 
         // Classify a sample text. Works only if trained.
-        int Classify(String sample);
+        // Result is in range [-1, 1] whereas -1 represents class 1 and 1 represents class 2.
+        double Classify(String sample);
     }
 
     public class NotTrainedException : System.Exception
